% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pollutants.R
\name{add_schwartz_pollutants}
\alias{add_schwartz_pollutants}
\title{add PM2.5, NO2, and O3 concentrations to data based on geohash}
\usage{
add_schwartz_pollutants(d, download_folder = fs::path_wd("s3_downloads"))
}
\arguments{
\item{d}{dataframe with columns called 'sitecode', 'start_date', and 'end_date'
(most likely the output from the `schwartz_grid_lookup`` container)}

\item{download_folder}{local path where files downloaded from s3 will be saved. Defaults to
a folder called 's3_downloads' in the current working directory.}
}
\value{
the input dataframe, expanded to include one row per day between the given 'start_date'
and 'end_date', with appended columns for geohash, PM2.5, NO2, and O3 concentrations.
}
\description{
add PM2.5, NO2, and O3 concentrations to data based on geohash
}
\examples{
if (FALSE) {
d <- tibble::tribble(
     ~id,         ~lat,    ~lon, ~site_index,      ~sitecode,  ~start_date,    ~end_date,
     '55000100280', 39.2, -84.6,   '9607238', '211050640897', '2008-09-09', '2008-09-11',
     '55000100281', 39.2, -84.6,   '9607238', '211050640897', '2007-08-05', '2007-08-08',
     '55000100282', 39.2, -84.6,   '9607238', '211050640897', '2015-08-31', '2015-09-02') \%>\%
   dplyr::mutate_at(vars(start_date, end_date), as.Date)

   add_schwartz_pollutants(d)
}
}
