% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_s3.R
\name{download_s3}
\alias{download_s3}
\title{easily download files from s3}
\usage{
download_s3(fl_names, s3_folder_url, download_dir)
}
\arguments{
\item{fl_names}{vector of file names (not the entire file path) to be downloaded from s3 to your local files}

\item{s3_folder_url}{the file location within s3, e.g. 's3://bucket_name/folder_name/'}

\item{download_dir}{the local path where the downloaded files are saved}
}
\value{
files downloaded to the specified directory
}
\description{
easily download files from s3
}
\examples{
if (FALSE) {
download_s3(fl_names =  c('dng_2016_round1.qs', 'dng_2015_round2.qs'),
            s3_folder_url = 's3://geomarker/schwartz/exp_estimates_1km/by_gh3_year/',
            download_dir = getwd())
}
}
